refactored websocket code to move to its own file
